+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and what APIs have changed, if applicable.
 - Add fluent client api for simple resource and association resource.
 - Add support for generating projection mask as the mask data map.
 - Fix UnmodifiableList wrap in d2 relative load balancer
+- Update fluent client APIs to include projection mask as input parameter.
 
 ## [29.16.0] - 2021-03-10
 - Add a ParSeq based CompletionStage implementation